Add wrap around protection
When creating the PIO encoder, there was an issue where rather than setting the device to negative values when turned, it would wrap to 2**32, cause the PID loop to massively overshoot in the wrong direction. A new class wraps this encoder so that one can get the position without this wrapping function. The method get_raw_pos() will get the unmodified version of this value in case the original function was intended.

Also, an invert option was added as well.